move all the components dealing with the location details into src/app-pages/map/component/location-details
