--- src/app-common/location-detail/location-detail.scss
@@ -1,19 +1,0 @@
-@import "../../style/sass/abstracts/variables";
-
-.location-detail-header {
-  width: 100%;
-  padding: 1em;
-  background-color: $base-lighter-gray;
-  .location-expand {
-    float:right;
-  }
-}
-
-.fixed-sticky {
-  position: -webkit-sticky; 
-  position: sticky;
-  top: 0;
-  z-index: 11;
-}
-
-
