--- src/app-common/map-details-container/map-details-content.scss
@@ -1,47 +1,0 @@
-@import "../../style/sass/abstracts/variables";
-@import "../../style/sass/abstracts/mixins";
-
-.location-detail-content-container {
-  display: flex;
-  flex-direction: row;
-  padding: 2rem;
-  min-height: 60vh;
-  @include mq($breakpoint: "phone", $direction: "down") {
-    flex-wrap: wrap;
-  }
-
-  .map-details-nav {
-    position: relative;
-    flex-shrink: 8;
-    width: 100%;
-    .fixed-sticky-nav {
-      position: -webkit-sticky;
-      position: sticky;
-      z-index: 10;
-      @include mq($breakpoint: "phone", $direction: "down") {
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-      }
-    }
-    .navbar-ul {
-      list-style: none;
-      padding-left: 0;
-      .nav-item {
-        button {
-          text-align: left;
-          padding-left: .25rem;
-          color: $black;
-        }
-      }
-      .nav-link-btn {
-        @include btn-to-link;
-        padding: 0.25rem 3rem;
-      }
-    }
-  }
-
-  .accordion-container {
-    width: 100%;
-  }
-}
