+++ src/app-pages/map/components/location-details/components/DamProfile.js
@@ -1,15 +1,14 @@
 import React, { useEffect,useState } from "react";
 import PropTypes from 'prop-types';
-import Table from "../../../app-common/table/Table";
 import { connect } from "redux-bundler-react";
-
+import Table from "../../../../../app-common/table/Table";
 
 const damProfileHeaderArr = ["name", "value"];
 let damProfileRowsArr;
 
 const DamProfile = ({ locationDetailData }) => {
 
-const [damProfileState, setDamProfileState] = useState( [] );
+  const [damProfileState, setDamProfileState] = useState( [] );
 
   useEffect (() => {
     if( locationDetailData.dam_profile && Array.isArray( locationDetailData.dam_profile.history ) ) {
@@ -17,15 +16,19 @@ const [damProfileState, setDamProfileState] = useState( [] );
       damProfileRowsArr = Object.entries(damProfileJsonObj);
       setDamProfileState(damProfileRowsArr);
     }
-  },[locationDetailData]);
+  }, [locationDetailData]);
 
   return (
     <div className="dam-profile-wrapper">
       <h5>Dam Profile Data</h5>
-    {damProfileState ? <Table rowsArr={damProfileState} headerRowArr={damProfileHeaderArr} /> :null}
+      { damProfileState
+        ? <Table rowsArr={damProfileState} headerRowArr={damProfileHeaderArr} />
+        : null
+      }
     </div>
   );
 };
+
 DamProfile.propTypes = {
   locationDetailData: PropTypes.object,
 };
