+++ src/app-pages/map/components/location-details/components/location-stream-controls/locationStreamControls.scss
@@ -1,4 +1,4 @@
-@import "../../../../style/sass/abstracts/mixins";
+@import "../../../../../../style/sass/abstracts/mixins";
 
 .location-stream-control-container {
   @include mq($breakpoint: "phone", $direction: "down") {
