+++ src/app-pages/map/components/location-details/components/time-series/TimeSeriesTable.js
@@ -1,6 +1,6 @@
 import React from "react";
 import PropTypes from "prop-types";
-import Sparkline from "../plotly/Sparkline";
+import Sparkline from "../../../../../../app-common/plotly//Sparkline";
 import "./time-series.scss";
 
 const TimeSeriesTable = ({
