+++ src/app-pages/map/components/location-details/components/time-series/time-series.scss
